requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
face_recognition
numpy>=1.20.0
opencv-python
dlib>=19.22.0
pillow>=10.0.0
python-jose[cryptography]
python-dotenv
pydantic 